services/camera-manager/requirements.txt
python-onvif-zeep==0.3.0
wsdiscovery==2.0.0
aiohttp==3.9.3
asyncpg==0.29.0
redis[hiredis]==5.0.1
pyyaml==6.0.1
structlog==24.1.0
cryptography==42.0.2
